refactor: avatar base64 conversion
